refactor(frontend): Separate `Home` page into separate components
