+++ Cargo.toml
@@ -27,7 +27,7 @@ serde = { version = "1.0.171", features = ["derive"] }
 anyhow = "1.0.71"
 once_cell = "1.18.0"
 serde_json = "1.0.100"
-tracing = { version = "0.1.38", features = ["release_max_level_debug"] }
+tracing = { version = "0.1.37", features = ["release_max_level_debug"] }
 tracing-subscriber = "0.3.17"
 hyper = { version = "0.14.27", features = ["server", "http1", "tcp"] }
 tokio-tungstenite = "0.19.0"
